sort out things and push to azure repos
