+++ src/Bot/HandlerResponse.cs
@@ -1,3 +1,3 @@
-﻿namespace Bot.Handlers;
+﻿namespace Bot;
 
 public record HandlerResponse<T>(T Result);
