+++ src/Bot/Nomination.cs
@@ -2,7 +2,6 @@ using System.Net;
 using System.Threading.Tasks;
 using Bot.Domain;
 using Bot.Dtos;
-using Bot.Handlers;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Logging;
@@ -11,21 +10,31 @@ namespace Bot;
 
 public class Nomination
 {
-    private readonly NominationHandler _nominationHandler;
+    private readonly IPersistence _persistence;
+    private readonly Messenger _messenger;
     private readonly ILogger _logger;
 
     public Nomination(
         ILoggerFactory loggerFactory,
-        NominationHandler nominationHandler)
+        IPersistence persistence,
+        Messenger messenger)
     {
-        _nominationHandler = nominationHandler;
+        _persistence = persistence;
+        _messenger = messenger;
         _logger = loggerFactory.CreateLogger<Nomination>();
     }
 
     [Function("Nomination_Timer")]
-    public async Task<Notification> NominationTimer([TimerTrigger("0 */5 * * * *")] ScheduleDto myTimer)
+    public async Task<Notification> NominationTimer(
+        [TimerTrigger("0 */5 * * * *"
+#if DEBUG
+        ,RunOnStartup = true
+#endif
+        )] object myTimer)
     {
-        var result = await _nominationHandler.Handle();
+        _ = myTimer;
+
+        var result = await Handle();
 
         return result.Result;
     }
@@ -36,12 +45,26 @@ public class Nomination
     {
         _logger.LogInformation("C# HTTP trigger function processed a request.");
 
-        var result = await _nominationHandler.Handle();
+        var result = await Handle();
 
         var response = req.CreateResponse(HttpStatusCode.OK);
 
         await response.WriteAsJsonAsync(result);
 
         return response;
+    }
+
+
+    public async Task<HandlerResponse<Notification>> Handle()
+    {
+        var boardMaster = await BoardMaster.Get(_persistence);
+
+        var nominee = boardMaster.Pick();
+
+        var notification = _messenger.Nudge(nominee);
+
+        await boardMaster.Save(_persistence);
+
+        return new HandlerResponse<Notification>(notification);
     }
 }
